# apps/backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/events/v2": {
      "post": {
        "operationId": "createEventV2",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDtoV2"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSingleResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create an event - V2",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events": {
      "post": {
        "operationId": "createEvent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSingleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid event payload"
          },
          "401": {
            "description": "No active session"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "500": {
            "description": "Event was not created"
          }
        },
        "summary": "Create an event",
        "tags": [
          "Events"
        ]
      },
      "get": {
        "description": "Apply filters - Enable stats mode with filters.stats=TRUE",
        "operationId": "getAllEvents",
        "parameters": [
          {
            "name": "moduleId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "timetableId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "all",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Stats",
            "required": false,
            "in": "query",
            "description": "Enable stats = TRUE",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponseDtoV2"
                }
              }
            }
          },
          "401": {
            "description": "No active session"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "summary": "Get all events",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{eventId}": {
      "get": {
        "operationId": "getEventById",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSingleResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event not found"
          }
        },
        "summary": "Get event by ID",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{id}": {
      "patch": {
        "operationId": "updateEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSingleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid update payload"
          },
          "401": {
            "description": "No active session"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Event or referenced module not found"
          },
          "500": {
            "description": "Event was not updated"
          }
        },
        "summary": "Update an event",
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "operationId": "deleteEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No active session"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Event not found"
          },
          "500": {
            "description": "Event was not deleted"
          }
        },
        "summary": "Delete an event",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/validate/{id}": {
      "patch": {
        "operationId": "validateEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event[${updated.eventName}] validated=[${updated.validated}]",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateEventResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Event not found"
          }
        },
        "summary": "Validate an event",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{id}/venue": {
      "patch": {
        "operationId": "updateEventVenue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventVenueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Venue updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSingleResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "summary": "Attach or clear an event's venue",
        "tags": [
          "Events"
        ]
      }
    },
...
    "/api/builder/events": {
      "post": {
        "description": "Create a new event, to module specified or personal module",
        "operationId": "BuilderController_createEvent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuilderEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSingleResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Module not found"
          }
        },
        "summary": "Create a user defined event",
        "tags": [
          "Builder"
        ]
      }
    },
...
    "/api/api-service/events": {
      "get": {
        "description": "Fetches events for a module at the authenticated user's university.",
        "operationId": "ApiServiceController_getEvents",
        "parameters": [
          {
            "name": "moduleId",
            "required": true,
            "in": "query",
            "description": "UUID of the module to fetch events for.",
            "schema": {
              "example": "00000000-0000-0000-0000-000000000000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The university or module ID is invalid."
          },
          "404": {
            "description": "The university, module, or API adapter could not be found."
          }
        },
        "summary": "Fetch events",
        "tags": [
          "ApiService"
        ]
      }
    }
  },
  "info": {
    "title": "UMTAS API",
    "description": "University Management Timetable & Scheduling System API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://capstone-vigil.dns.net.za",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "umtas-session",
        "description": "Session cookie (set automatically by BetterAuth)"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EventCriteriaDto": {
        "type": "object",
        "properties": {
          "eventSource": {
            "enum": [
              "university",
              "personal"
            ],
            "type": "string"
          },
          "date": {
            "type": "string",
            "example": "2026-02-17",
            "description": "Required when the event is not recurring."
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "monday",
              "tuesday",
              "wednesday",
              "thursday",
              "friday",
              "saturday",
              "sunday"
            ],
            "description": "Required when the event is recurring."
          },
          "startTime": {
            "type": "string",
            "example": "08:30"
          },
          "endTime": {
            "type": "string",
            "example": "10:20"
          },
          "moduleId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "eventSource",
          "startTime",
          "endTime"
        ]
      },
      "VenueDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "venueName": {
            "type": "string"
          }
        },
        "required": [
          "venueId",
          "venueName"
        ]
      },
      "EventDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "eventName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "activityCode": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 10
          },
          "activityType": {
            "type": "string",
            "enum": [
              "lecture",
              "tutorial",
              "prac",
              "test",
              "exam"
            ]
          },
          "eventCriteria": {
            "$ref": "#/components/schemas/EventCriteriaDto"
          },
          "isRecurring": {
            "type": "boolean"
          },
          "validated": {
            "type": "boolean"
          },
          "venues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueDto"
            }
          }
        },
        "required": [
          "eventId",
          "eventCriteria"
        ]
      },
...
      "EventCriteriaDtoV2": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-02-17",
            "description": "Required when the event is not recurring."
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "monday",
              "tuesday",
              "wednesday",
              "thursday",
              "friday",
              "saturday",
              "sunday"
            ],
            "description": "Required when the event is recurring."
          },
          "startTime": {
            "type": "string",
            "example": "08:30"
          },
          "endTime": {
            "type": "string",
            "example": "10:20"
          },
          "moduleId": {
            "type": "string",
            "format": "uuid",
            "default": false,
            "description": "moduleId required in V2 - whomp whomp",
            "nullable": false
          }
        }
      },
      "CreateEventDtoV2": {
        "type": "object",
        "properties": {
          "eventName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "activityCode": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 10
          },
          "activityType": {
            "type": "string",
            "enum": [
              "lecture",
              "tutorial",
              "prac",
              "test",
              "exam"
            ]
          },
          "isRecurring": {
            "type": "boolean"
          },
          "validated": {
            "type": "boolean"
          },
          "venues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueDto"
            }
          },
          "eventCriteria": {
            "$ref": "#/components/schemas/EventCriteriaDtoV2"
          }
        },
        "required": [
          "eventCriteria"
        ]
      },
      "EventSingleResponseDto": {
        "type": "object",
        "properties": {
          "event": {
            "$ref": "#/components/schemas/EventDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "event"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "eventName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "activityCode": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 10
          },
          "eventCriteria": {
            "$ref": "#/components/schemas/EventCriteriaDto"
          },
          "isRecurring": {
            "type": "boolean"
          },
          "validated": {
            "type": "boolean"
          },
          "venues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueDto"
            }
          },
          "activityType": {
            "type": "string",
            "enum": [
              "lecture",
              "tutorial",
              "prac",
              "test",
              "exam"
            ],
            "description": "Required when eventCriteria.moduleId is provided."
          }
        },
        "required": [
          "eventCriteria"
        ]
      },
      "EventListResponseDtoV2": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          },
          "count": {
            "type": "number",
            "example": 68,
            "description": "Count of Modules"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "events"
        ]
      },
      "UpdateEventCriteriaDto": {
        "type": "object",
        "properties": {
          "eventSource": {
            "enum": [
              "university",
              "personal"
            ],
            "type": "string"
          },
          "date": {
            "type": "string",
            "example": "2026-02-17",
            "description": "Required when the event is not recurring."
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "monday",
              "tuesday",
              "wednesday",
              "thursday",
              "friday",
              "saturday",
              "sunday"
            ],
            "description": "Required when the event is recurring."
          },
          "startTime": {
            "type": "string",
            "example": "08:30"
          },
          "endTime": {
            "type": "string",
            "example": "10:20"
          },
          "moduleId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "eventName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "activityCode": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 10
          },
          "activityType": {
            "type": "string",
            "enum": [
              "lecture",
              "tutorial",
              "prac",
              "test",
              "exam"
            ]
          },
          "isRecurring": {
            "type": "boolean"
          },
          "validated": {
            "type": "boolean"
          },
          "eventCriteria": {
            "$ref": "#/components/schemas/UpdateEventCriteriaDto"
          }
        }
      },
      "DeleteResponseDto": {
        "type": "object",
        "properties": {
          "eventName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "activityCode": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 10
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "ValidateEventDto": {
        "type": "object",
        "properties": {
          "validated": {
            "type": "boolean"
          }
        }
      },
      "ValidateEventResponseDto": {
        "type": "object",
        "properties": {
          "event": {
            "$ref": "#/components/schemas/EventDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "event"
        ]
      },
      "UpdateEventVenueDto": {
        "type": "object",
        "properties": {
          "venueName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "The venue name"
          },
          "buildingId": {
            "type": "string",
            "format": "uuid",
            "description": "The selected buildingdD"
          }
        },
        "required": [
          "venueName"
        ]
      },
...
      "CreateBuilderEventDto": {
        "type": "object",
        "properties": {
          "eventName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "activityCode": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 10
          },
          "eventCriteria": {
            "$ref": "#/components/schemas/EventCriteriaDto"
          },
          "isRecurring": {
            "type": "boolean"
          },
          "activityType": {
            "type": "string",
            "enum": [
              "lecture",
              "tutorial",
              "prac",
              "test",
              "exam"
            ],
            "description": "Required when eventCriteria.moduleId is provided."
          }
        }
      },
...
      "EventListResponseDto": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "events"
        ]
      }
    }
  }
}